bot de telegram/4.2/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt --resolver=backtracking requirements.in
#
anyio==3.6.2
    # via httpcore
certifi==2022.12.7
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.0.1
    # via requests
h11==0.14.0
    # via httpcore
h2==4.1.0
    # via httpx
hpack==4.0.0
    # via h2
httpcore==0.16.3
    # via httpx
httpx[http2]==0.23.3
    # via python-telegram-bot
hyperframe==6.0.1
    # via h2
idna==3.4
    # via
    #   anyio
    #   requests
    #   rfc3986
python-telegram-bot==20.1
    # via -r requirements.in
requests==2.28.2
    # via -r requirements.in
rfc3986[idna2008]==1.5.0
    # via httpx
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
urllib3==1.26.14
    # via requests
